refactor/godClass-ParamsBuildHelper
