+++ src/main/java/com/ly/doc/helper/ParamsBuildHelper.java
@@ -55,6 +55,7 @@ import com.ly.doc.model.DocJavaField;
 import com.ly.doc.utils.DocClassUtil;
 import com.ly.doc.utils.JavaFieldUtil;
 import com.ly.doc.utils.ParamUtil;
+import com.ly.doc.utils.ParamsBuildHelperUtil;
 import com.thoughtworks.qdox.model.JavaAnnotation;
 import com.thoughtworks.qdox.model.JavaClass;
 import com.thoughtworks.qdox.model.JavaField;
@@ -112,7 +113,7 @@ public class ParamsBuildHelper extends BaseHelper {
         List<DocJavaField> fields = JavaClassUtil.getFields(cls, 0, new LinkedHashMap<>(),projectBuilder.getApiConfig().getClassLoader());
         if (JavaClassValidateUtil.isPrimitive(simpleName)) {
             String processedType = isShowJavaType ? simpleName : DocClassUtil.processTypeNameForParams(simpleName.toLowerCase());
-            paramList.addAll(primitiveReturnRespComment(processedType, atomicInteger, pid));
+            paramList.addAll(ParamsBuildHelperUtil.primitiveReturnRespComment(processedType, atomicInteger, pid));
         } else if (JavaClassValidateUtil.isCollection(simpleName) || JavaClassValidateUtil.isArray(simpleName)) {
             if (!JavaClassValidateUtil.isCollection(globGicName[0])) {
                 String gNameTemp = globGicName[0];
@@ -120,7 +121,7 @@ public class ParamsBuildHelper extends BaseHelper {
                 if (JavaClassValidateUtil.isPrimitive(gName)) {
                     String processedType = isShowJavaType ? simpleName : DocClassUtil.processTypeNameForParams(gName);
                     ApiParam param = ApiParam.of()
-                            .setId(atomicOrDefault(atomicInteger, pid + 1))
+                            .setId(ParamsBuildHelperUtil.atomicOrDefault(atomicInteger, pid + 1))
                             .setField(pre + " -")
                             .setType("array[" + processedType + "]")
                             .setPid(pid)
@@ -137,12 +138,12 @@ public class ParamsBuildHelper extends BaseHelper {
                 }
             }
         } else if (JavaClassValidateUtil.isMap(simpleName)) {
-            paramList.addAll(buildMapParam(globGicName, pre, level, isRequired, isResp,
+            paramList.addAll(ParamsBuildHelperUtil.buildMapParam(globGicName, pre, level, isRequired, isResp,
                     registryClasses, projectBuilder, groupClasses, pid, jsonRequest, nextLevel, atomicInteger));
         } else if (DocGlobalConstants.JAVA_OBJECT_FULLY.equals(className)) {
             ApiParam param = ApiParam.of()
                     .setClassName(className)
-                    .setId(atomicOrDefault(atomicInteger, pid + 1))
+                    .setId(ParamsBuildHelperUtil.atomicOrDefault(atomicInteger, pid + 1))
                     .setField(pre + "any object")
                     .setType("object")
                     .setPid(pid)
@@ -290,7 +291,7 @@ public class ParamsBuildHelper extends BaseHelper {
                     ApiParam param = ApiParam.of().setField(pre + fieldName).setType("file")
                             .setClassName(className)
                             .setPid(pid)
-                            .setId(atomicOrDefault(atomicInteger, paramList.size() + pid + 1))
+                            .setId(ParamsBuildHelperUtil.atomicOrDefault(atomicInteger, paramList.size() + pid + 1))
                             .setMaxLength(maxLength)
                             .setDesc(comment.toString())
                             .setRequired(strRequired)
@@ -309,11 +310,11 @@ public class ParamsBuildHelper extends BaseHelper {
                     }
                     ApiParam param = ApiParam.of().setClassName(className).setField(pre + fieldName);
                     param.setPid(pid).setMaxLength(maxLength).setValue(fieldValue);
-                    param.setId(atomicOrDefault(atomicInteger, paramList.size() + param.getPid() + 1));
+                    param.setId(ParamsBuildHelperUtil.atomicOrDefault(atomicInteger, paramList.size() + param.getPid() + 1));
                     String processedType = isShowJavaType ? subTypeName : DocClassUtil.processTypeNameForParams(subTypeName.toLowerCase());
                     param.setType(processedType);
                     // handle param
-                    commonHandleParam(paramList, param, isRequired, comment.toString(), since, strRequired);
+                    ParamsBuildHelperUtil.commonHandleParam(paramList, param, isRequired, comment.toString(), since, strRequired);
 
                     JavaClass enumClass = ParamUtil.handleSeeEnum(param, field, projectBuilder, jsonRequest, tagsMap);
                     if (Objects.nonNull(enumClass)) {
@@ -322,7 +323,7 @@ public class ParamsBuildHelper extends BaseHelper {
                             enumClassComment = enumClass.getComment();
                         }
                         comment = new StringBuilder(StringUtils.isEmpty(comment.toString()) ? enumClassComment : comment.toString());
-                        String enumComment = handleEnumComment(enumClass, projectBuilder);
+                        String enumComment = ParamsBuildHelperUtil.handleEnumComment(enumClass, projectBuilder);
                         param.setDesc(comment + enumComment);
                     }
                 } else {
@@ -346,7 +347,7 @@ public class ParamsBuildHelper extends BaseHelper {
                     preBuilder.append("└─");
                     int fieldPid;
                     ApiParam param = ApiParam.of().setField(pre + fieldName).setClassName(className).setPid(pid).setMaxLength(maxLength);
-                    param.setId(atomicOrDefault(atomicInteger, paramList.size() + param.getPid() + 1));
+                    param.setId(ParamsBuildHelperUtil.atomicOrDefault(atomicInteger, paramList.size() + param.getPid() + 1));
                     String processedType;
                     if (fieldGicName.length() == 1) {
                         String gicName = DocGlobalConstants.JAVA_OBJECT_FULLY;
@@ -368,10 +369,10 @@ public class ParamsBuildHelper extends BaseHelper {
                     param.setType(processedType);
                     JavaClass javaClass = field.getType();
                     if (javaClass.isEnum()) {
-                        comment.append(handleEnumComment(javaClass, projectBuilder));
+                        comment.append(ParamsBuildHelperUtil.handleEnumComment(javaClass, projectBuilder));
                         ParamUtil.handleSeeEnum(param, field, projectBuilder, jsonRequest, tagsMap);
                         // hand Param
-                        commonHandleParam(paramList, param, isRequired, comment + appendComment, since, strRequired);
+                        ParamsBuildHelperUtil.commonHandleParam(paramList, param, isRequired, comment + appendComment, since, strRequired);
                     } else if (JavaClassValidateUtil.isCollection(subTypeName) || JavaClassValidateUtil.isArray(subTypeName)) {
                         if (isShowJavaType) {
                             // rpc
@@ -397,7 +398,7 @@ public class ParamsBuildHelper extends BaseHelper {
                         if (gNameArr.length > 0) {
                             String gName = DocClassUtil.getSimpleGicName(fieldGicName)[0];
                             JavaClass javaClass1 = projectBuilder.getJavaProjectBuilder().getClassByName(gName);
-                            comment.append(handleEnumComment(javaClass1, projectBuilder));
+                            comment.append(ParamsBuildHelperUtil.handleEnumComment(javaClass1, projectBuilder));
                         }
                         String gName = gNameArr[0];
                         if (JavaClassValidateUtil.isPrimitive(gName)) {
@@ -408,9 +409,9 @@ public class ParamsBuildHelper extends BaseHelper {
                             } else {
                                 param.setValue(fieldValue);
                             }
-                            commonHandleParam(paramList, param, isRequired, comment + appendComment, since, strRequired);
+                            ParamsBuildHelperUtil.commonHandleParam(paramList, param, isRequired, comment + appendComment, since, strRequired);
                         } else {
-                            commonHandleParam(paramList, param, isRequired, comment + appendComment, since, strRequired);
+                            ParamsBuildHelperUtil.commonHandleParam(paramList, param, isRequired, comment + appendComment, since, strRequired);
                             fieldPid = Optional.ofNullable(atomicInteger).isPresent() ? param.getId() : paramList.size() + pid;
                             if (!simpleName.equals(gName)) {
                                 JavaClass arraySubClass = projectBuilder.getJavaProjectBuilder().getClassByName(gName);
@@ -445,14 +446,14 @@ public class ParamsBuildHelper extends BaseHelper {
                             param.setType("map");
                             param.setValue(fieldValue);
                         }
-                        commonHandleParam(paramList, param, isRequired, comment + appendComment, since, strRequired);
+                        ParamsBuildHelperUtil.commonHandleParam(paramList, param, isRequired, comment + appendComment, since, strRequired);
                         fieldPid = Optional.ofNullable(atomicInteger).isPresent() ? param.getId() : paramList.size() + pid;
                         String valType = DocClassUtil.getMapKeyValueType(fieldGicName).length == 0 ? fieldGicName
                                 : DocClassUtil.getMapKeyValueType(fieldGicName)[1];
                         if (JavaClassValidateUtil.isMap(fieldGicName) || DocGlobalConstants.JAVA_OBJECT_FULLY.equals(valType)) {
                             ApiParam param1 = ApiParam.of()
                                     .setField(preBuilder.toString() + "any object")
-                                    .setId(atomicOrDefault(atomicInteger, fieldPid + 1)).setPid(fieldPid)
+                                    .setId(ParamsBuildHelperUtil.atomicOrDefault(atomicInteger, fieldPid + 1)).setPid(fieldPid)
                                     .setClassName(className)
                                     .setMaxLength(maxLength)
                                     .setType("object")
@@ -477,9 +478,9 @@ public class ParamsBuildHelper extends BaseHelper {
                         if (StringUtil.isEmpty(param.getDesc())) {
                             param.setDesc(DocGlobalConstants.ANY_OBJECT_MSG);
                         }
-                        commonHandleParam(paramList, param, isRequired, comment + appendComment, since, strRequired);
+                        ParamsBuildHelperUtil.commonHandleParam(paramList, param, isRequired, comment + appendComment, since, strRequired);
                     } else if (fieldGicName.length() == 1) {
-                        commonHandleParam(paramList, param, isRequired, comment + appendComment, since, strRequired);
+                        ParamsBuildHelperUtil.commonHandleParam(paramList, param, isRequired, comment + appendComment, since, strRequired);
                         fieldPid = Optional.ofNullable(atomicInteger).isPresent() ? param.getId() : paramList.size() + pid;
                         // handle java generic or object
                         if (!simpleName.equals(className)) {
@@ -518,7 +519,7 @@ public class ParamsBuildHelper extends BaseHelper {
                         ApiParam param1 = ApiParam.of()
                                 .setField(pre + fieldName)
                                 .setPid(pid)
-                                .setId(atomicOrDefault(atomicInteger, paramList.size() + pid + 1))
+                                .setId(ParamsBuildHelperUtil.atomicOrDefault(atomicInteger, paramList.size() + pid + 1))
                                 .setClassName(subTypeName)
                                 .setMaxLength(maxLength)
                                 .setType("object")
@@ -526,7 +527,7 @@ public class ParamsBuildHelper extends BaseHelper {
                                 .setVersion(DocGlobalConstants.DEFAULT_VERSION);
                         paramList.add(param1);
                     } else {
-                        commonHandleParam(paramList, param, isRequired, comment + appendComment, since, strRequired);
+                        ParamsBuildHelperUtil.commonHandleParam(paramList, param, isRequired, comment + appendComment, since, strRequired);
                         fieldGicName = DocUtil.formatFieldTypeGicName(genericMap, globGicName, fieldGicName);
                         fieldPid = Optional.ofNullable(atomicInteger).isPresent() ? param.getId() : paramList.size() + pid;
                         paramList.addAll(buildParams(fieldGicName, preBuilder.toString(), nextLevel, isRequired
@@ -537,113 +538,5 @@ public class ParamsBuildHelper extends BaseHelper {
             }//end field
         }
         return paramList;
-    }
-
-    private static List<ApiParam> buildMapParam(String[] globGicName, String pre, int level, String isRequired, boolean isResp,
-                                                Map<String, String> registryClasses,
-                                                ProjectDocConfigBuilder projectBuilder, Set<String> groupClasses, int pid, boolean jsonRequest,
-                                                int nextLevel, AtomicInteger atomicInteger) {
-        if (globGicName.length != 2) {
-            return Collections.emptyList();
-        }
-
-        // mock map key param
-        String mapKeySimpleName = DocClassUtil.getSimpleName(globGicName[0]);
-        String valueSimpleName = DocClassUtil.getSimpleName(globGicName[1]);
-
-        List<ApiParam> paramList = new ArrayList<>();
-        if (JavaClassValidateUtil.isPrimitive(mapKeySimpleName)) {
-            boolean isShowJavaType = projectBuilder.getApiConfig().getShowJavaType();
-            String valueSimpleNameType = isShowJavaType ? valueSimpleName : DocClassUtil.processTypeNameForParams(valueSimpleName.toLowerCase());
-            ApiParam apiParam = ApiParam.of().setField(pre + "mapKey")
-                    .setType(valueSimpleNameType)
-                    .setClassName(valueSimpleName)
-                    .setDesc(Optional.ofNullable(projectBuilder.getClassByName(valueSimpleName)).map(JavaClass::getComment).orElse("A map key."))
-                    .setVersion(DocGlobalConstants.DEFAULT_VERSION)
-                    .setPid(pid)
-                    .setId(atomicOrDefault(atomicInteger, ++pid));
-            paramList.add(apiParam);
-        }
-        // build param when map value is not primitive
-        if (JavaClassValidateUtil.isPrimitive(valueSimpleName)) {
-            return paramList;
-        }
-        StringBuilder preBuilder = new StringBuilder();
-        for (int j = 0; j < level; j++) {
-            preBuilder.append(DocGlobalConstants.FIELD_SPACE);
-        }
-        preBuilder.append("└─");
-        paramList.addAll(buildParams(globGicName[1], preBuilder.toString(), ++nextLevel, isRequired, isResp
-                , registryClasses, projectBuilder, groupClasses, pid, jsonRequest, atomicInteger));
-        return paramList;
-    }
-
-    public static String dictionaryListComment(List<EnumDictionary> enumDataDict) {
-        return enumDataDict.stream().map(apiDataDictionary ->
-                apiDataDictionary.getName() + "-(\"" + apiDataDictionary.getValue() + "\",\"" + apiDataDictionary.getDesc() + "\")"
-        ).collect(Collectors.joining(","));
-    }
-
-    public static List<ApiParam> primitiveReturnRespComment(String typeName, AtomicInteger atomicInteger, int pid) {
-        String comments = "Return " + typeName + ".";
-        ApiParam apiParam = ApiParam.of().setClassName(typeName)
-                .setId(atomicOrDefault(atomicInteger, pid + 1))
-                .setField("-")
-                .setPid(pid)
-                .setType(typeName)
-                .setDesc(comments)
-                .setVersion(DocGlobalConstants.DEFAULT_VERSION);
-
-        List<ApiParam> paramList = new ArrayList<>();
-        paramList.add(apiParam);
-        return paramList;
-    }
-
-    private static void commonHandleParam(List<ApiParam> paramList, ApiParam param, String isRequired
-            , String comment, String since, boolean strRequired) {
-        if (StringUtil.isEmpty(isRequired)) {
-            param.setDesc(comment).setVersion(since);
-        } else {
-            param.setDesc(comment).setVersion(since).setRequired(strRequired);
-        }
-        paramList.add(param);
-    }
-
-    private static String handleEnumComment(JavaClass javaClass, ProjectDocConfigBuilder projectBuilder) {
-        String comment = "";
-        if (!javaClass.isEnum()) {
-            return comment;
-        }
-        String enumComments = javaClass.getComment();
-        if (Boolean.TRUE.equals(projectBuilder.getApiConfig().getInlineEnum())) {
-            ApiDataDictionary dataDictionary = projectBuilder.getApiConfig().getDataDictionary(javaClass.getCanonicalName());
-            if (Objects.isNull(dataDictionary)) {
-                comment = comment + "<br/>[Enum values:<br/>" + JavaClassUtil.getEnumParams(javaClass) + "]";
-            } else {
-                Class enumClass = dataDictionary.getEnumClass();
-                if (enumClass.isInterface()) {
-                    ClassLoader classLoader = projectBuilder.getApiConfig().getClassLoader();
-                    try {
-                        enumClass = classLoader.loadClass(javaClass.getFullyQualifiedName());
-                    } catch (ClassNotFoundException e) {
-                        return comment;
-                    }
-                }
-                comment = comment + "<br/>[Enum:" + dictionaryListComment(dataDictionary.getEnumDataDict(enumClass)) + "]";
-            }
-        } else {
-            if (StringUtil.isNotEmpty(enumComments)) {
-                comment = comment + "<br/>(See: " + enumComments + ")";
-            }
-            comment = StringUtil.removeQuotes(comment);
-        }
-        return comment;
-    }
-
-    private static int atomicOrDefault(AtomicInteger atomicInteger, int defaultVal) {
-        if (null != atomicInteger) {
-            return atomicInteger.incrementAndGet();
-        }
-        return defaultVal;
     }
 }
